+++ NEWS.md
@@ -3,6 +3,8 @@
 * `uncount()` error for non-integer weights and gives a clearer error message
   for negative weights (@mgirlich, #1069).
 
+* You can once again unnest dates (#1021, #1089).
+
 * `pivot_wider()` works with data.table and empty key variables (@mgirlich, #1066).
 
 * `separate_rows()` works for factor columns (@mgirlich, #1058).